requirements.txt
fastapi
uvicorn[standard]
python-dotenv    
requests
pydantic
pandas
numpy
scikit-learn
joblib
python-dateutil
APScheduler
python-jose[cryptography]
passlib[bcrypt]
python-multipart